Add Sha256 Check and Creat a specific father directory for each item.

+++ main.py
@@ -1,3 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+import hashlib
 import json
 import os
 
@@ -47,36 +50,53 @@ def season_acquire(month: str):
     return season
 
 
-def markdown_generator(obj: Animation):
+def markdown_generator(jsonfile: str):
+    obj = json_phase(jsonfile)
+    hashcode = gethash(jsonfile)
     date = obj.start.split('.')
     season = season_acquire(date[1])
-    dir = './database/' + date[0] + '/' + season + '/'
+    dir = './database/' + date[0] + '/' + season + '/' + obj.name + '/'
     filename = dir + obj.name + '.md'
-    print(filename)
+    hashfile = dir + hashcode + '.json'
     if not os.path.exists(dir):
         os.makedirs(dir)
-    file = open(filename, 'w', encoding='utf-8')
-    file.write("# " + obj.name + '\n\n')
-    file.write("|------|スタッフ|\n")
-    file.write("|:------:|:------:|\n")
-    file.write("|原作|" + obj.author + '|\n')
-    file.write("|原作イラスト|" + obj.illustrator + '|\n')
-    file.write("|監督|" + obj.director + '|\n')
-    file.write("|アニメーション制作|" + obj.studio + '|\n')
-    file.write("|製作委員会|" + obj.committee + '|\n')
-    file.write("|放送開始期間|" + obj.start + '|\n')
-    file.write("|放送終了期間|" + obj.end + '|\n')
-    file.write("|話数|" + str(obj.episode) + '|\n')
-    file.write("|OVA話数|" + str(obj.ova) + '|\n')
-    file.close()
+    if os.path.exists(hashfile):
+        print(obj.name + " is not Changed, skip...")
+        return None
+    print('Generate copy and markdown for ' + obj.name + '...')
+    hashfile = open(hashfile, 'w', encoding='utf-8')
+    hashfile.write(json_to_str(jsonfile).encode('utf8').decode('unicode_escape'))
+    hashfile.close()
+    mdfile = open(filename, 'w', encoding='utf-8')
+    mdfile.write("# " + obj.name + '\n\n')
+    mdfile.write("|------|スタッフ|\n")
+    mdfile.write("|:------:|:------:|\n")
+    mdfile.write("|原作|" + obj.author + '|\n')
+    mdfile.write("|原作イラスト|" + obj.illustrator + '|\n')
+    mdfile.write("|監督|" + obj.director + '|\n')
+    mdfile.write("|アニメーション制作|" + obj.studio + '|\n')
+    mdfile.write("|製作委員会|" + obj.committee + '|\n')
+    mdfile.write("|放送開始期間|" + obj.start + '|\n')
+    mdfile.write("|放送終了期間|" + obj.end + '|\n')
+    mdfile.write("|話数|" + str(obj.episode) + '|\n')
+    mdfile.write("|OVA話数|" + str(obj.ova) + '|\n')
+    mdfile.close()
 
 
-def process():
+def process() -> None:
     files = os.listdir(path='./_temp')
     for file in files:
         if file.endswith('.json'):
             filename = './_temp/' + file
-            markdown_generator(json_phase(filename))
+            markdown_generator(filename)
+
+
+def json_to_str(jsonfile: str) -> str:
+    return json.dumps(json.loads(open(jsonfile, 'r', encoding='utf-8').read()))
+
+
+def gethash(jsonfile: str) -> str:
+    return hashlib.sha256(json_to_str(jsonfile).encode()).hexdigest()
 
 
 if __name__ == '__main__':
